\name{compo.cocktail}

\alias{compo.cocktail}

\docType{data}

\title{Composition of the cocktails data}

\description{
The data used here refer to the composition of 16 cocktails, i.e. the mango, banana, orange and lemon concentration.
}

\usage{data(cocktail)}

\format{
A data frame with 16 rows and 4 columns: the composition of each cocktail is given for the 4 ingredients.
}

\source{
Département de mathématiques appliquées, Agrocampus Rennes
}

\examples{
data(cocktail)
}

\keyword{datasets}
