\name{hedo.cocktail}

\alias{hedo.cocktail}

\docType{data}

\title{Cocktails hedonic scores}

\description{
The data used here refer to 16 cocktails.
Each cocktail was evaluated on a structured scale from 0 to 10, 
by 100 consumers, according to their liking (0) or disliking (10).
}

\usage{data(cocktail)}

\format{
 A data frame with 16 rows and 100 columns: each row corresponds to a cocktail
 and each column to the hedonic scores given by one of the 100 consumers participating in the study.
}

\source{
Département de mathématiques appliquées, Agrocampus Rennes
}

\examples{
data(cocktail)
}

\keyword{datasets}
