\name{hedochoc}

\alias{hedochoc}

\docType{data}

\title{Chocolates hedonic scores}

\description{
The data used here refer to six varieties of chocolates sold in France.
Each chocolate was evaluated on a structured scale from 0 to 10, 
by 222 consumers, according to their liking (0) or disliking (10).
}

\usage{data(chocolates)}

\format{
 A data frame with 6 rows and 222 columns: each row corresponds to a chocolate
 and each column to the hedonic scores given by one of the 222 consumers participating in the study.
}

\source{
Département de mathématiques appliquées, Agrocampus Rennes
}

\examples{
data(chocolates)
}

\keyword{datasets}
