\name{plotellipseinter}
\alias{plotellipseinter}

\title{Plotellipseinter}
\description{Construit l'ensemble des ellipses de confiance}
\usage{
plotellipseinter(mat,alpha=0.05,coord=c(1,2),nbbloc=1,moy=TRUE)
}

\arguments{
  \item{mat}{Matrix with all the points}
  \item{coord}{a length 2 vector specifying the components to plot}
  \item{alpha}{Confidence level of the ellipses}
  \item{nbbloc}{Confidence level of the ellipses}
  \item{moy}{Confidence level of the ellipses}
}

\author{François Husson}

\keyword{internal}
