\name{senso.cocktail}

\alias{senso.cocktail}

\docType{data}

\title{Sensory data for 16 cocktails}

\description{
The data used here refer to the sensory description of 16 cocktails. Each cocktail was evaluated by 12 panelists according to 13
sensory descriptors (only the average of each cocktail are given).
}

\usage{data(cocktail)}

\format{
A data frame with 16 rows and 13 columns: each cocktail was evaluated by 12 panelists according to 13
sensory descriptors.
}

\source{
Département de mathématiques appliquées, Agrocampus Rennes
}

\examples{
data(cocktail)
}

\keyword{datasets}
