\name{senso.scatterutil.eti.circ}
\alias{senso.scatterutil.eti.circ}

\title{senso.scatterutil.eti.circ}
\description{
Puts labels on a correlation circle
}
\usage{
senso.scatterutil.eti.circ(x, y,label,clabel, origin = c(0, 0), boxes = FALSE, col="black")
}
\arguments{
  \item{x}{}
  \item{y}{}
  \item{label}{}
  \item{clabel}{}
  \item{origin}{}
  \item{boxes}{}
}


\author{
Delphine Anger
}

\keyword{internal}
