\name{senso.scatterutil.eti}

\alias{senso.scatterutil.eti}

\title{senso.scatterutil.eti}

\description{
\code{scatter} is a generic function. It has methods for the classes \code{coa}, \code{dudi}, \code{fca}, \code{acm} and \code{pco}.\cr
The scale of the grid is situated on the right-top of the graph.\cr
The points are in the middle of the labels.\cr
This process plots the graphs of the multivariate analyses.\cr
The two axes have the same scale.\cr
}

\usage{
senso.scatterutil.eti(x, y, label, clabel, boxes = FALSE, 
    coul = rep(1, length(x)), font = 2, pos = 4, offset = 0.2)
}
\arguments{
  \item{x}{}
  \item{y}{}
  \item{label}{}
  \item{clabel}{}
  \item{boxes}{}
  \item{coul}{}
  \item{font}{}
  \item{pos}{}
  \item{offset}{}
}

\author{Daniel Chessel \email{chessel@biomserv.univ-lyon1.fr}}

\keyword{internal}
