\name{sensopanels}

\alias{sensopanels}

\docType{data}

\title{Sensory profiles given by 7 panels}

\description{
The data used here refer to six varieties of chocolates sold in France.
Each chocolate was evaluated by 7 panels according to 14
sensory descriptors.
}

\usage{data(chocolates)}

\format{
 A data frame with 6 rows and 98 columns: each row corresponds to a chocolate
 and each column to the mean over the panelists of a given panel 
 according to a sensory descriptor.
}

\source{
Département de mathématiques appliquées, Agrocampus Rennes
}

\examples{
data(chocolates)
}

\keyword{datasets}
