\name{WilliamsDesign}

\alias{WilliamsDesign}

\title{Construct a Williams design}

\description{
Construct a Williams design
}

\usage{
WilliamsDesign(nbprod,seed=NULL)
}

\arguments{
  \item{nbprod}{nombre de produits}
  \item{seed}{a number to intialize the generator of the random number}
}

\value{
Retrun the design\cr
}

\author{Francois Husson \email{husson@agrocampus-rennes.fr}}

\examples{
 exemple1 <- WilliamsDesign(12)
}

\keyword{models}
